requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
pypdf==4.3.1
python-dotenv==1.0.1
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.1.0
faiss-cpu